requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
selenium>=4.15.0
webdriver-manager>=4.0.0
pymongo>=4.6.0
matplotlib>=3.7.0
seaborn>=0.12.0
numpy>=1.24.0
yfinance>=0.2.0
scikit-learn>=1.3.0
scipy>=1.11.0
polygon-api-client>=1.13.0
streamlit>=1.28.0
python-dotenv>=1.0.0

